refactor(onboard): extract isGatewayTcpReady into a focused module
Addresses PR review feedback — onboard.ts is the God Object being
decomposed, and growing its exports is an anti-pattern. Follow the
pattern established by PR #3312 (which added
src/lib/onboard/gateway-http-readiness.ts) and create a peer module
for the Docker-driver path.

New file: src/lib/onboard/gateway-tcp-readiness.ts
  Exports: isGatewayTcpReady(port, timeoutMs, host)
  Model: mirrors gateway-http-readiness.ts structure exactly —
         default port from GATEWAY_PORT, minimum-timeout clamp,
         pure async, never throws.

New file: src/lib/onboard/gateway-tcp-readiness.test.ts
  Co-located behavioural tests per the established pattern
  (see temp-files.test.ts, legacy-command.test.ts, etc.)
  6 tests: listening / closed / timeout / min-timeout / no-throw /
  default-port. Imports via ES module, not via dist/ indirection.

onboard.ts net change:
  Before: +88 lines (+48 helper + integration)
  After:  +22 lines (2 require lines + integration only)

Related to: #3111
